* flutter_master: (810 commits)
  Revert engine roll to 316b026807c50d65ba1b6724ac954711683f1b01
  roll engine to 316b026807c50d65ba1b6724ac954711683f1b01 (#19419)
  Revert "enable lint unnecessary_const (#19342)" (#19423)
  enable lint unnecessary_const (#19342)
  Chevrons in month picker are semi-transparent when the month is scrolled (#19363)
  Revert "Use FlutterProject to locate files  (#18913)" (#19409)
  Extra debug information in run_release_test (#19405)
  Fix typo (#19402)
  Use FlutterProject to locate files  (#18913)
  Revert "roll engine to 9af920ef22f311588b649ba876127993f5c1362c (#19365)" (#19376)
  roll engine to 9af920ef22f311588b649ba876127993f5c1362c (#19365)
  increase cache size if image is loaded that is larger than max size (#19352)
  Add Bash and Zsh command-line completion for flutter (#19243)
  Support keyboardAppearance field for iOS (#19244)
  Add option to silence driver extension errors (#19247)
  Add HeroController to CupertinoApp (#19326)
  have text finder convert Text.rich to plain text for comparison (#19270)
  Correct contentPadding type in InputDecoration.debugFillProperties (#19318)
  Revert "Revert "Use runTests in fuchsia tester. (#19178)" (#19321)" (#19327)
  Roll engine to c5a63d (#19288)
  ...

+++ dev/devicelab/lib/tasks/microbenchmarks.dart
@@ -14,7 +14,7 @@ import 'package:flutter_devicelab/framework/ios.dart';
 import 'package:flutter_devicelab/framework/utils.dart';
 
 /// The maximum amount of time a single microbenchmark is allowed to take.
-const Duration _kBenchmarkTimeout = const Duration(minutes: 6);
+const Duration _kBenchmarkTimeout = const Duration(minutes: 10);
 
 /// Creates a device lab task that runs benchmarks in
 /// `dev/benchmarks/microbenchmarks` reports results to the dashboard.
@@ -23,7 +23,7 @@ TaskFunction createMicrobenchmarkTask() {
     final Device device = await devices.workingDevice;
     await device.unlock();
 
-    Future<Map<String, double>> _runMicrobench(String benchmarkPath, {bool previewDart2: true}) async {
+    Future<Map<String, double>> _runMicrobench(String benchmarkPath, {bool previewDart2 = true}) async {
       Future<Map<String, double>> _run() async {
         print('Running $benchmarkPath');
         final Directory appDir = dir(
@@ -74,8 +74,6 @@ TaskFunction createMicrobenchmarkTask() {
     addDart1Results(await _runMicrobench(
         'lib/stocks/layout_bench.dart', previewDart2: false));
     addDart1Results(await _runMicrobench(
-        'lib/stocks/layout_bench.dart', previewDart2: false));
-    addDart1Results(await _runMicrobench(
         'lib/stocks/build_bench.dart', previewDart2: false));
     addDart1Results(await _runMicrobench(
         'lib/gestures/velocity_tracker_bench.dart', previewDart2: false));
@@ -86,9 +84,9 @@ TaskFunction createMicrobenchmarkTask() {
 }
 
 Future<Process> _startFlutter({
-  String command: 'run',
-  List<String> options: const <String>[],
-  bool canFail: false,
+  String command = 'run',
+  List<String> options = const <String>[],
+  bool canFail = false,
   Map<String, String> environment,
 }) {
   final List<String> args = <String>['run']..addAll(options);
@@ -112,6 +110,7 @@ Future<Map<String, double>> _readJsonResults(Process process) {
       });
 
   bool processWasKilledIntentionally = false;
+  bool resultsHaveBeenParsed = false;
   final StreamSubscription<String> stdoutSub = process.stdout
       .transform(const Utf8Decoder())
       .transform(const LineSplitter())
@@ -124,20 +123,44 @@ Future<Map<String, double>> _readJsonResults(Process process) {
     }
 
     if (line.contains(jsonEnd)) {
+      final String jsonOutput = jsonBuf.toString();
+
+      // If we end up here and have already parsed the results, it suggests that
+      // we have recieved output from another test because our `flutter run`
+      // process did not terminate correctly.
+      // https://github.com/flutter/flutter/issues/19096#issuecomment-402756549
+      if (resultsHaveBeenParsed) {
+        throw 'Additional JSON was received after results has already been '
+              'processed. This suggests the `flutter run` process may have lived '
+              'past the end of our test and collected additional output from the '
+              'next test.\n\n'
+              'The JSON below contains all collected output, including both from '
+              'the original test and what followed.\n\n'
+              '$jsonOutput';
+      }
+
       jsonStarted = false;
       processWasKilledIntentionally = true;
+      resultsHaveBeenParsed = true;
+      // ignore: deprecated_member_use
       process.kill(ProcessSignal.SIGINT); // flutter run doesn't quit automatically
-      completer.complete(json.decode(jsonBuf.toString()));
+      try {
+        completer.complete(new Map<String, double>.from(json.decode(jsonOutput)));
+      } catch (ex) {
+        completer.completeError('Decoding JSON failed ($ex). JSON string was: $jsonOutput');
+      }
       return;
     }
 
-    if (jsonStarted)
+    if (jsonStarted && line.contains(jsonPrefix))
       jsonBuf.writeln(line.substring(line.indexOf(jsonPrefix) + jsonPrefix.length));
   });
 
-  process.exitCode.then<int>((int code) {
-    stdoutSub.cancel();
-    stderrSub.cancel();
+  process.exitCode.then<int>((int code) async {
+    await Future.wait<void>(<Future<void>>[
+      stdoutSub.cancel(),
+      stderrSub.cancel(),
+    ]);
     if (!processWasKilledIntentionally && code != 0) {
       completer.completeError('flutter run failed: exit code=$code');
     }
